+++ src/app/event-handlers.ts
@@ -1,0 +1,685 @@
+import type { AppContext, AppModule } from '@/app/app-context';
+import type { PanelConfig } from '@/types';
+import type { MapView } from '@/components';
+import type { ClusteredEvent } from '@/types';
+import type { DashboardSnapshot } from '@/services/storage';
+import {
+  PlaybackControl,
+  StatusPanel,
+  MobileWarningModal,
+  PizzIntIndicator,
+  CIIPanel,
+  PredictionPanel,
+} from '@/components';
+import {
+  buildMapUrl,
+  debounce,
+  saveToStorage,
+  ExportPanel,
+  getCurrentTheme,
+  setTheme,
+} from '@/utils';
+import {
+  STORAGE_KEYS,
+  SITE_VARIANT,
+  LAYER_TO_SOURCE,
+  FEEDS,
+  INTEL_SOURCES,
+} from '@/config';
+import {
+  saveSnapshot,
+  initAisStream,
+  disconnectAisStream,
+} from '@/services';
+import {
+  trackPanelView,
+  trackVariantSwitch,
+  trackThemeChanged,
+  trackMapViewChange,
+  trackMapLayerToggle,
+  trackPanelToggled,
+} from '@/services/analytics';
+import { invokeTauri } from '@/services/tauri-bridge';
+import { dataFreshness } from '@/services/data-freshness';
+import { mlWorker } from '@/services/ml-worker';
+import { UnifiedSettings } from '@/components/UnifiedSettings';
+import { t } from '@/services/i18n';
+
+export interface EventHandlerCallbacks {
+  updateSearchIndex: () => void;
+  loadAllData: () => Promise<void>;
+  flushStaleRefreshes: () => void;
+  setHiddenSince: (ts: number) => void;
+  loadDataForLayer: (layer: string) => void;
+  waitForAisData: () => void;
+  syncDataFreshnessWithLayers: () => void;
+}
+
+export class EventHandlerManager implements AppModule {
+  private ctx: AppContext;
+  private callbacks: EventHandlerCallbacks;
+
+  private boundFullscreenHandler: (() => void) | null = null;
+  private boundResizeHandler: (() => void) | null = null;
+  private boundVisibilityHandler: (() => void) | null = null;
+  private boundDesktopExternalLinkHandler: ((e: MouseEvent) => void) | null = null;
+  private boundIdleResetHandler: (() => void) | null = null;
+  private idleTimeoutId: ReturnType<typeof setTimeout> | null = null;
+  private snapshotIntervalId: ReturnType<typeof setInterval> | null = null;
+  private clockIntervalId: ReturnType<typeof setInterval> | null = null;
+  private readonly IDLE_PAUSE_MS = 2 * 60 * 1000;
+
+  constructor(ctx: AppContext, callbacks: EventHandlerCallbacks) {
+    this.ctx = ctx;
+    this.callbacks = callbacks;
+  }
+
+  init(): void {
+    this.setupEventListeners();
+    this.setupIdleDetection();
+  }
+
+  destroy(): void {
+    if (this.boundFullscreenHandler) {
+      document.removeEventListener('fullscreenchange', this.boundFullscreenHandler);
+      this.boundFullscreenHandler = null;
+    }
+    if (this.boundResizeHandler) {
+      window.removeEventListener('resize', this.boundResizeHandler);
+      this.boundResizeHandler = null;
+    }
+    if (this.boundVisibilityHandler) {
+      document.removeEventListener('visibilitychange', this.boundVisibilityHandler);
+      this.boundVisibilityHandler = null;
+    }
+    if (this.boundDesktopExternalLinkHandler) {
+      document.removeEventListener('click', this.boundDesktopExternalLinkHandler, true);
+      this.boundDesktopExternalLinkHandler = null;
+    }
+    if (this.idleTimeoutId) {
+      clearTimeout(this.idleTimeoutId);
+      this.idleTimeoutId = null;
+    }
+    if (this.boundIdleResetHandler) {
+      ['mousedown', 'keydown', 'scroll', 'touchstart', 'mousemove'].forEach(event => {
+        document.removeEventListener(event, this.boundIdleResetHandler!);
+      });
+      this.boundIdleResetHandler = null;
+    }
+    if (this.snapshotIntervalId) {
+      clearInterval(this.snapshotIntervalId);
+      this.snapshotIntervalId = null;
+    }
+    if (this.clockIntervalId) {
+      clearInterval(this.clockIntervalId);
+      this.clockIntervalId = null;
+    }
+    this.ctx.unifiedSettings?.destroy();
+    this.ctx.unifiedSettings = null;
+  }
+
+  private setupEventListeners(): void {
+    document.getElementById('searchBtn')?.addEventListener('click', () => {
+      this.callbacks.updateSearchIndex();
+      this.ctx.searchModal?.open();
+    });
+
+    document.getElementById('copyLinkBtn')?.addEventListener('click', async () => {
+      const shareUrl = this.getShareUrl();
+      if (!shareUrl) return;
+      const button = document.getElementById('copyLinkBtn');
+      try {
+        await this.copyToClipboard(shareUrl);
+        this.setCopyLinkFeedback(button, 'Copied!');
+      } catch (error) {
+        console.warn('Failed to copy share link:', error);
+        this.setCopyLinkFeedback(button, 'Copy failed');
+      }
+    });
+
+    window.addEventListener('storage', (e) => {
+      if (e.key === STORAGE_KEYS.panels && e.newValue) {
+        try {
+          this.ctx.panelSettings = JSON.parse(e.newValue) as Record<string, PanelConfig>;
+          this.applyPanelSettings();
+          this.ctx.unifiedSettings?.refreshPanelToggles();
+        } catch (_) {}
+      }
+      if (e.key === STORAGE_KEYS.liveChannels && e.newValue) {
+        const panel = this.ctx.panels['live-news'];
+        if (panel && typeof (panel as unknown as { refreshChannelsFromStorage?: () => void }).refreshChannelsFromStorage === 'function') {
+          (panel as unknown as { refreshChannelsFromStorage: () => void }).refreshChannelsFromStorage();
+        }
+      }
+    });
+
+    document.getElementById('headerThemeToggle')?.addEventListener('click', () => {
+      const next = getCurrentTheme() === 'dark' ? 'light' : 'dark';
+      setTheme(next);
+      this.updateHeaderThemeIcon();
+      trackThemeChanged(next);
+    });
+
+    const isLocalDev = location.hostname === 'localhost' || location.hostname === '127.0.0.1';
+    if (this.ctx.isDesktopApp || isLocalDev) {
+      this.ctx.container.querySelectorAll<HTMLAnchorElement>('.variant-option').forEach(link => {
+        link.addEventListener('click', (e) => {
+          const variant = link.dataset.variant;
+          if (variant && variant !== SITE_VARIANT) {
+            e.preventDefault();
+            trackVariantSwitch(SITE_VARIANT, variant);
+            localStorage.setItem('worldmonitor-variant', variant);
+            window.location.reload();
+          }
+        });
+      });
+    }
+
+    const fullscreenBtn = document.getElementById('fullscreenBtn');
+    if (!this.ctx.isDesktopApp && fullscreenBtn) {
+      fullscreenBtn.addEventListener('click', () => this.toggleFullscreen());
+      this.boundFullscreenHandler = () => {
+        fullscreenBtn.textContent = document.fullscreenElement ? '\u26F6' : '\u26F6';
+        fullscreenBtn.classList.toggle('active', !!document.fullscreenElement);
+      };
+      document.addEventListener('fullscreenchange', this.boundFullscreenHandler);
+    }
+
+    const regionSelect = document.getElementById('regionSelect') as HTMLSelectElement;
+    regionSelect?.addEventListener('change', () => {
+      this.ctx.map?.setView(regionSelect.value as MapView);
+      trackMapViewChange(regionSelect.value);
+    });
+
+    this.boundResizeHandler = () => {
+      this.ctx.map?.render();
+    };
+    window.addEventListener('resize', this.boundResizeHandler);
+
+    this.setupMapResize();
+    this.setupMapPin();
+
+    this.boundVisibilityHandler = () => {
+      document.body.classList.toggle('animations-paused', document.hidden);
+      if (document.hidden) {
+        this.callbacks.setHiddenSince(Date.now());
+        mlWorker.unloadOptionalModels();
+      } else {
+        this.resetIdleTimer();
+        this.callbacks.flushStaleRefreshes();
+      }
+    };
+    document.addEventListener('visibilitychange', this.boundVisibilityHandler);
+
+    window.addEventListener('focal-points-ready', () => {
+      (this.ctx.panels['cii'] as CIIPanel)?.refresh(true);
+    });
+
+    window.addEventListener('theme-changed', () => {
+      this.ctx.map?.render();
+      this.updateHeaderThemeIcon();
+    });
+
+    if (this.ctx.isDesktopApp) {
+      if (this.boundDesktopExternalLinkHandler) {
+        document.removeEventListener('click', this.boundDesktopExternalLinkHandler, true);
+      }
+      this.boundDesktopExternalLinkHandler = (e: MouseEvent) => {
+        if (!(e.target instanceof Element)) return;
+        const anchor = e.target.closest('a[href]') as HTMLAnchorElement | null;
+        if (!anchor) return;
+        const href = anchor.href;
+        if (!href || href.startsWith('javascript:') || href === '#' || href.startsWith('#')) return;
+        try {
+          const url = new URL(href, window.location.href);
+          if (url.origin === window.location.origin) return;
+          e.preventDefault();
+          e.stopPropagation();
+          void invokeTauri<void>('open_url', { url: url.toString() }).catch(() => {
+            window.open(url.toString(), '_blank');
+          });
+        } catch { /* malformed URL -- let browser handle */ }
+      };
+      document.addEventListener('click', this.boundDesktopExternalLinkHandler, true);
+    }
+  }
+
+  private setupIdleDetection(): void {
+    this.boundIdleResetHandler = () => {
+      if (this.ctx.isIdle) {
+        this.ctx.isIdle = false;
+        document.body.classList.remove('animations-paused');
+      }
+      this.resetIdleTimer();
+    };
+
+    ['mousedown', 'keydown', 'scroll', 'touchstart', 'mousemove'].forEach(event => {
+      document.addEventListener(event, this.boundIdleResetHandler!, { passive: true });
+    });
+
+    this.resetIdleTimer();
+  }
+
+  resetIdleTimer(): void {
+    if (this.idleTimeoutId) {
+      clearTimeout(this.idleTimeoutId);
+    }
+    this.idleTimeoutId = setTimeout(() => {
+      if (!document.hidden) {
+        this.ctx.isIdle = true;
+        document.body.classList.add('animations-paused');
+        console.log('[App] User idle - pausing animations to save resources');
+      }
+    }, this.IDLE_PAUSE_MS);
+  }
+
+  setupUrlStateSync(): void {
+    if (!this.ctx.map) return;
+    const update = debounce(() => {
+      const shareUrl = this.getShareUrl();
+      if (!shareUrl) return;
+      history.replaceState(null, '', shareUrl);
+    }, 250);
+
+    this.ctx.map.onStateChanged(() => {
+      update();
+      const regionSelect = document.getElementById('regionSelect') as HTMLSelectElement;
+      if (regionSelect && this.ctx.map) {
+        const state = this.ctx.map.getState();
+        if (regionSelect.value !== state.view) {
+          regionSelect.value = state.view;
+        }
+      }
+    });
+    update();
+  }
+
+  getShareUrl(): string | null {
+    if (!this.ctx.map) return null;
+    const state = this.ctx.map.getState();
+    const center = this.ctx.map.getCenter();
+    const baseUrl = `${window.location.origin}${window.location.pathname}`;
+    return buildMapUrl(baseUrl, {
+      view: state.view,
+      zoom: state.zoom,
+      center,
+      timeRange: state.timeRange,
+      layers: state.layers,
+      country: this.ctx.countryBriefPage?.isVisible() ? (this.ctx.countryBriefPage.getCode() ?? undefined) : undefined,
+    });
+  }
+
+  private async copyToClipboard(text: string): Promise<void> {
+    if (navigator.clipboard?.writeText) {
+      await navigator.clipboard.writeText(text);
+      return;
+    }
+    const textarea = document.createElement('textarea');
+    textarea.value = text;
+    textarea.style.position = 'fixed';
+    textarea.style.opacity = '0';
+    document.body.appendChild(textarea);
+    textarea.select();
+    document.execCommand('copy');
+    document.body.removeChild(textarea);
+  }
+
+  private setCopyLinkFeedback(button: HTMLElement | null, message: string): void {
+    if (!button) return;
+    const originalText = button.textContent ?? '';
+    button.textContent = message;
+    button.classList.add('copied');
+    window.setTimeout(() => {
+      button.textContent = originalText;
+      button.classList.remove('copied');
+    }, 1500);
+  }
+
+  toggleFullscreen(): void {
+    if (document.fullscreenElement) {
+      try { void document.exitFullscreen()?.catch(() => {}); } catch {}
+    } else {
+      const el = document.documentElement as HTMLElement & { webkitRequestFullscreen?: () => void };
+      if (el.requestFullscreen) {
+        try { void el.requestFullscreen()?.catch(() => {}); } catch {}
+      } else if (el.webkitRequestFullscreen) {
+        try { el.webkitRequestFullscreen(); } catch {}
+      }
+    }
+  }
+
+  updateHeaderThemeIcon(): void {
+    const btn = document.getElementById('headerThemeToggle');
+    if (!btn) return;
+    const isDark = getCurrentTheme() === 'dark';
+    btn.innerHTML = isDark
+      ? '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><circle cx="12" cy="12" r="5"/><line x1="12" y1="1" x2="12" y2="3"/><line x1="12" y1="21" x2="12" y2="23"/><line x1="4.22" y1="4.22" x2="5.64" y2="5.64"/><line x1="18.36" y1="18.36" x2="19.78" y2="19.78"/><line x1="1" y1="12" x2="3" y2="12"/><line x1="21" y1="12" x2="23" y2="12"/><line x1="4.22" y1="19.78" x2="5.64" y2="18.36"/><line x1="18.36" y1="5.64" x2="19.78" y2="4.22"/></svg>'
+      : '<svg width="16" height="16" viewBox="0 0 24 24" fill="none" stroke="currentColor" stroke-width="2" stroke-linecap="round" stroke-linejoin="round"><path d="M21 12.79A9 9 0 1111.21 3 7 7 0 0021 12.79z"/></svg>';
+  }
+
+  startHeaderClock(): void {
+    const el = document.getElementById('headerClock');
+    if (!el) return;
+    const tick = () => {
+      el.textContent = new Date().toUTCString().replace('GMT', 'UTC');
+    };
+    tick();
+    this.clockIntervalId = setInterval(tick, 1000);
+  }
+
+  setupMobileWarning(): void {
+    if (MobileWarningModal.shouldShow()) {
+      this.ctx.mobileWarningModal = new MobileWarningModal();
+      this.ctx.mobileWarningModal.show();
+    }
+  }
+
+  setupStatusPanel(): void {
+    this.ctx.statusPanel = new StatusPanel();
+    const headerLeft = this.ctx.container.querySelector('.header-left');
+    if (headerLeft) {
+      headerLeft.appendChild(this.ctx.statusPanel.getElement());
+    }
+  }
+
+  setupPizzIntIndicator(): void {
+    if (SITE_VARIANT === 'tech' || SITE_VARIANT === 'finance') return;
+
+    this.ctx.pizzintIndicator = new PizzIntIndicator();
+    const headerLeft = this.ctx.container.querySelector('.header-left');
+    if (headerLeft) {
+      headerLeft.appendChild(this.ctx.pizzintIndicator.getElement());
+    }
+  }
+
+  setupExportPanel(): void {
+    this.ctx.exportPanel = new ExportPanel(() => ({
+      news: this.ctx.latestClusters.length > 0 ? this.ctx.latestClusters : this.ctx.allNews,
+      markets: this.ctx.latestMarkets,
+      predictions: this.ctx.latestPredictions,
+      timestamp: Date.now(),
+    }));
+
+    const headerRight = this.ctx.container.querySelector('.header-right');
+    if (headerRight) {
+      headerRight.insertBefore(this.ctx.exportPanel.getElement(), headerRight.firstChild);
+    }
+  }
+
+  setupUnifiedSettings(): void {
+    this.ctx.unifiedSettings = new UnifiedSettings({
+      getPanelSettings: () => this.ctx.panelSettings,
+      togglePanel: (key: string) => {
+        const config = this.ctx.panelSettings[key];
+        if (config) {
+          config.enabled = !config.enabled;
+          trackPanelToggled(key, config.enabled);
+          saveToStorage(STORAGE_KEYS.panels, this.ctx.panelSettings);
+          this.applyPanelSettings();
+        }
+      },
+      getDisabledSources: () => this.ctx.disabledSources,
+      toggleSource: (name: string) => {
+        if (this.ctx.disabledSources.has(name)) {
+          this.ctx.disabledSources.delete(name);
+        } else {
+          this.ctx.disabledSources.add(name);
+        }
+        saveToStorage(STORAGE_KEYS.disabledFeeds, Array.from(this.ctx.disabledSources));
+      },
+      setSourcesEnabled: (names: string[], enabled: boolean) => {
+        for (const name of names) {
+          if (enabled) this.ctx.disabledSources.delete(name);
+          else this.ctx.disabledSources.add(name);
+        }
+        saveToStorage(STORAGE_KEYS.disabledFeeds, Array.from(this.ctx.disabledSources));
+      },
+      getAllSourceNames: () => this.getAllSourceNames(),
+      getLocalizedPanelName: (key: string, fallback: string) => this.getLocalizedPanelName(key, fallback),
+      isDesktopApp: this.ctx.isDesktopApp,
+    });
+
+    const mount = document.getElementById('unifiedSettingsMount');
+    if (mount) {
+      mount.appendChild(this.ctx.unifiedSettings.getButton());
+    }
+  }
+
+  setupPlaybackControl(): void {
+    this.ctx.playbackControl = new PlaybackControl();
+    this.ctx.playbackControl.onSnapshot((snapshot) => {
+      if (snapshot) {
+        this.ctx.isPlaybackMode = true;
+        this.restoreSnapshot(snapshot);
+      } else {
+        this.ctx.isPlaybackMode = false;
+        this.callbacks.loadAllData();
+      }
+    });
+
+    const headerRight = this.ctx.container.querySelector('.header-right');
+    if (headerRight) {
+      headerRight.insertBefore(this.ctx.playbackControl.getElement(), headerRight.firstChild);
+    }
+  }
+
+  setupSnapshotSaving(): void {
+    const saveCurrentSnapshot = async () => {
+      if (this.ctx.isPlaybackMode || this.ctx.isDestroyed) return;
+
+      const marketPrices: Record<string, number> = {};
+      this.ctx.latestMarkets.forEach(m => {
+        if (m.price !== null) marketPrices[m.symbol] = m.price;
+      });
+
+      await saveSnapshot({
+        timestamp: Date.now(),
+        events: this.ctx.latestClusters,
+        marketPrices,
+        predictions: this.ctx.latestPredictions.map(p => ({
+          title: p.title,
+          yesPrice: p.yesPrice
+        })),
+        hotspotLevels: this.ctx.map?.getHotspotLevels() ?? {}
+      });
+    };
+
+    void saveCurrentSnapshot().catch((e) => console.warn('[Snapshot] save failed:', e));
+    this.snapshotIntervalId = setInterval(() => void saveCurrentSnapshot().catch((e) => console.warn('[Snapshot] save failed:', e)), 15 * 60 * 1000);
+  }
+
+  restoreSnapshot(snapshot: DashboardSnapshot): void {
+    for (const panel of Object.values(this.ctx.newsPanels)) {
+      panel.showLoading();
+    }
+
+    const events = snapshot.events as ClusteredEvent[];
+    this.ctx.latestClusters = events;
+
+    const predictions = snapshot.predictions.map((p, i) => ({
+      id: `snap-${i}`,
+      title: p.title,
+      yesPrice: p.yesPrice,
+      noPrice: 100 - p.yesPrice,
+      volume24h: 0,
+      liquidity: 0,
+    }));
+    this.ctx.latestPredictions = predictions;
+    (this.ctx.panels['polymarket'] as PredictionPanel).renderPredictions(predictions);
+
+    this.ctx.map?.setHotspotLevels(snapshot.hotspotLevels);
+  }
+
+  setupMapLayerHandlers(): void {
+    this.ctx.map?.setOnLayerChange((layer, enabled, source) => {
+      console.log(`[App.onLayerChange] ${layer}: ${enabled} (${source})`);
+      trackMapLayerToggle(layer, enabled, source);
+      this.ctx.mapLayers[layer] = enabled;
+      saveToStorage(STORAGE_KEYS.mapLayers, this.ctx.mapLayers);
+
+      const sourceIds = LAYER_TO_SOURCE[layer];
+      if (sourceIds) {
+        for (const sourceId of sourceIds) {
+          dataFreshness.setEnabled(sourceId, enabled);
+        }
+      }
+
+      if (layer === 'ais') {
+        if (enabled) {
+          this.ctx.map?.setLayerLoading('ais', true);
+          initAisStream();
+          this.callbacks.waitForAisData();
+        } else {
+          disconnectAisStream();
+        }
+        return;
+      }
+
+      if (enabled) {
+        this.callbacks.loadDataForLayer(layer);
+      }
+    });
+  }
+
+  setupPanelViewTracking(): void {
+    const viewedPanels = new Set<string>();
+    const observer = new IntersectionObserver((entries) => {
+      for (const entry of entries) {
+        if (entry.isIntersecting && entry.intersectionRatio >= 0.3) {
+          const id = (entry.target as HTMLElement).dataset.panel;
+          if (id && !viewedPanels.has(id)) {
+            viewedPanels.add(id);
+            trackPanelView(id);
+          }
+        }
+      }
+    }, { threshold: 0.3 });
+
+    const grid = document.getElementById('panelsGrid');
+    if (grid) {
+      for (const child of Array.from(grid.children)) {
+        if ((child as HTMLElement).dataset.panel) {
+          observer.observe(child);
+        }
+      }
+    }
+  }
+
+  showToast(msg: string): void {
+    document.querySelector('.toast-notification')?.remove();
+    const el = document.createElement('div');
+    el.className = 'toast-notification';
+    el.textContent = msg;
+    document.body.appendChild(el);
+    requestAnimationFrame(() => el.classList.add('visible'));
+    setTimeout(() => { el.classList.remove('visible'); setTimeout(() => el.remove(), 300); }, 3000);
+  }
+
+  shouldShowIntelligenceNotifications(): boolean {
+    return !this.ctx.isMobile && !!this.ctx.findingsBadge?.isPopupEnabled();
+  }
+
+  setupMapResize(): void {
+    const mapSection = document.getElementById('mapSection');
+    const resizeHandle = document.getElementById('mapResizeHandle');
+    if (!mapSection || !resizeHandle) return;
+
+    const getMinHeight = () => (window.innerWidth >= 2000 ? 320 : 400);
+    const getMaxHeight = () => Math.max(getMinHeight(), window.innerHeight - 60);
+
+    const savedHeight = localStorage.getItem('map-height');
+    if (savedHeight) {
+      const numeric = Number.parseInt(savedHeight, 10);
+      if (Number.isFinite(numeric)) {
+        const clamped = Math.max(getMinHeight(), Math.min(numeric, getMaxHeight()));
+        mapSection.style.height = `${clamped}px`;
+        if (clamped !== numeric) {
+          localStorage.setItem('map-height', `${clamped}px`);
+        }
+      } else {
+        localStorage.removeItem('map-height');
+      }
+    }
+
+    let isResizing = false;
+    let startY = 0;
+    let startHeight = 0;
+
+    resizeHandle.addEventListener('mousedown', (e) => {
+      isResizing = true;
+      startY = e.clientY;
+      startHeight = mapSection.offsetHeight;
+      mapSection.classList.add('resizing');
+      document.body.style.cursor = 'ns-resize';
+      e.preventDefault();
+    });
+
+    document.addEventListener('mousemove', (e) => {
+      if (!isResizing) return;
+      const deltaY = e.clientY - startY;
+      const newHeight = Math.max(getMinHeight(), Math.min(startHeight + deltaY, getMaxHeight()));
+      mapSection.style.height = `${newHeight}px`;
+      this.ctx.map?.render();
+    });
+
+    document.addEventListener('mouseup', () => {
+      if (!isResizing) return;
+      isResizing = false;
+      mapSection.classList.remove('resizing');
+      document.body.style.cursor = '';
+      localStorage.setItem('map-height', mapSection.style.height);
+      this.ctx.map?.render();
+    });
+  }
+
+  setupMapPin(): void {
+    const mapSection = document.getElementById('mapSection');
+    const pinBtn = document.getElementById('mapPinBtn');
+    if (!mapSection || !pinBtn) return;
+
+    const isPinned = localStorage.getItem('map-pinned') === 'true';
+    if (isPinned) {
+      mapSection.classList.add('pinned');
+      pinBtn.classList.add('active');
+    }
+
+    pinBtn.addEventListener('click', () => {
+      const nowPinned = mapSection.classList.toggle('pinned');
+      pinBtn.classList.toggle('active', nowPinned);
+      localStorage.setItem('map-pinned', String(nowPinned));
+    });
+  }
+
+  getLocalizedPanelName(panelKey: string, fallback: string): string {
+    if (panelKey === 'runtime-config') {
+      return t('modals.runtimeConfig.title');
+    }
+    const key = panelKey.replace(/-([a-z])/g, (_match, group: string) => group.toUpperCase());
+    const lookup = `panels.${key}`;
+    const localized = t(lookup);
+    return localized === lookup ? fallback : localized;
+  }
+
+  getAllSourceNames(): string[] {
+    const sources = new Set<string>();
+    Object.values(FEEDS).forEach(feeds => {
+      if (feeds) feeds.forEach(f => sources.add(f.name));
+    });
+    INTEL_SOURCES.forEach(f => sources.add(f.name));
+    return Array.from(sources).sort((a, b) => a.localeCompare(b));
+  }
+
+  applyPanelSettings(): void {
+    Object.entries(this.ctx.panelSettings).forEach(([key, config]) => {
+      if (key === 'map') {
+        const mapSection = document.getElementById('mapSection');
+        if (mapSection) {
+          mapSection.classList.toggle('hidden', !config.enabled);
+        }
+        return;
+      }
+      const panel = this.ctx.panels[key];
+      panel?.toggle(config.enabled);
+    });
+  }
+}
