+++ src/app/index.ts
@@ -1,0 +1,8 @@
+export type { AppContext, AppModule, CountryBriefSignals, IntelligenceCache } from './app-context';
+export { DesktopUpdater } from './desktop-updater';
+export { CountryIntelManager } from './country-intel';
+export { SearchManager } from './search-manager';
+export { RefreshScheduler } from './refresh-scheduler';
+export { PanelLayoutManager } from './panel-layout';
+export { DataLoaderManager } from './data-loader';
+export { EventHandlerManager } from './event-handlers';
